exception handling

+++ src/main/java/com/system/recipeblog/services/IngredientServiceImpl.java
@@ -1,5 +1,7 @@
 package com.system.recipeblog.services;
 
+import com.system.recipeblog.exception.ex.IngredientNotFoundException;
+import com.system.recipeblog.exception.ex.RecipeNotFoundException;
 import com.system.recipeblog.models.Ingredient;
 import com.system.recipeblog.models.Recipe;
 import com.system.recipeblog.repositories.IngredientRepository;
@@ -11,6 +13,9 @@ import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
+import static com.system.recipeblog.util.Constants.INGREDIENT_NOT_FOUND;
+import static com.system.recipeblog.util.Constants.RECIPE_NOT_FOUND;
+
 @Service
 @Transactional
 @RequiredArgsConstructor
@@ -19,15 +24,10 @@ public class IngredientServiceImpl implements IngredientService{
     private final RecipeRepository recipeRepository;
 
     @Override
-    public Ingredient saveOrUpdateIngredient(Ingredient ingredient, Recipe recipe) {
-        Optional<Recipe> recipeOptional = recipeRepository.findById(recipe.getId());
+    public Ingredient saveOrUpdateIngredient(Ingredient ingredient, Long recipeId) {
+        Recipe recipe = recipeRepository.findById(recipeId)
+                .orElseThrow(() -> new RecipeNotFoundException(String.format(RECIPE_NOT_FOUND, recipeId)));
 
-        if (!recipeOptional.isPresent()) {
-            System.out.println("Recipe not found");
-            return null; // Or throw an exception if necessary
-        }
-
-        recipe = recipeOptional.get();
 
         Optional<Ingredient> ingredientOptional = recipe.getIngredients().stream().findFirst();
 
@@ -40,7 +40,7 @@ public class IngredientServiceImpl implements IngredientService{
             recipe.addIngredient(ingredient);
         }
 
-        Recipe savedRecipe = recipeRepository.save(recipe);
+        recipeRepository.save(recipe);
         return ingredientRepository.save(ingredient);
     }
 
@@ -50,12 +50,16 @@ public class IngredientServiceImpl implements IngredientService{
     }
 
     @Override
-    public Optional<Ingredient> findById(Long id) {
-        return ingredientRepository.findById(id);
+    public Ingredient findById(Long id) {
+        return ingredientRepository.findById(id)
+                .orElseThrow(() -> new IngredientNotFoundException(String.format(INGREDIENT_NOT_FOUND, id)));
     }
 
     @Override
     public void deleteIngredient(Long id) {
+        if (!ingredientRepository.existsById(id)) {
+            throw new IngredientNotFoundException(String.format(INGREDIENT_NOT_FOUND, id));
+        }
         ingredientRepository.deleteById(id);
     }
 }
